1. Created classes Dog and Cat with inheritance from the class Animal.
2. Animals can perform actions: run, swim, jump over an obstacle.
As a parameter, each method is given a value indicating either
the length of the obstacle (for running and swimming)
or the height (for jumping).

3. Each animal has restrictions on actions
(running: cat 200 m, dog 500 m, jump: cat 2 m, dog 0.5 m,
swimming: cat can not swim, dog 10 m).

4. When an animal attempts to perform one of these actions,
it must report the result to the console.

5. Animals have the scatter in the constraints.
That is, one dog has a run limit of 400 m, another 600 m.